Changing the apiCounter type (from INT to BIGINT)

+++ src/main/java/com/zeniapp/segmentmiddleware/entities/Session.java
@@ -35,7 +35,7 @@ public class Session {
     @Getter @Setter
     private Account account;
 
-    @Column(name = "api_counter", columnDefinition = "INT", updatable = true)
+    @Column(name = "api_counter", columnDefinition = "BIGINT", updatable = true)
     @Setter
     @Getter
     private Integer apiCounter;
